Tests actually have run-hooks.sh on PATH
I had thought they did before, since most of the tests were passing.

I was entirely wrong. $0 is *not* the name of the test suite in
bats (and there is no reason on earth why I should have expected it to
be, since bats is the running binary).

This commit stands as a testament to the difficulty of writing good
tests. Just checking for the expected exit status of 1 turned out to be
woefully insufficient.