Source package's githooks.conf.sh file
This gives hook packages a way to customize the environment they run in.

This should let you isolate packages cleanly from one another, but the
current implementation probably doesn't do that correctly.

Now that I have a working test for the behavior, though, I should be
able to add further tests to verify that each package runs in a clean
environment.